Progress.  Debugging up to but not including hello.java(34).

+++ src/net/three4j/core/Uint16BufferAttribute.java
@@ -1,19 +1,23 @@
 package net.three4j.core;
 
-import org.mozilla.types.Int8Array;
 import org.mozilla.types.Uint16Array;
 
 public class Uint16BufferAttribute extends BufferAttribute {
 
-	public Uint16BufferAttribute(Int8Array array, int itemSize, boolean normalized) {
+	public Uint16BufferAttribute(Uint16Array array, int itemSize, boolean normalized) {
 	}
 
-	public Uint16BufferAttribute(double[] _array, int i) {
+	public Uint16BufferAttribute(int[] _array, int i) {
 		this(_array, i, false);
 	}
 
-	public Uint16BufferAttribute(double[] _array, int i, boolean normalized) {
+	public Uint16BufferAttribute(int[] _array, int i, boolean normalized) {
 		super(new Uint16Array(_array), i, normalized);
 	}
+
+//	@Override
+//	public int getX(int i) {
+//		return 1;
+//	}
 
 }
